Optimisation
Don't do anything with the file if it is not compressed.

Previously the file would be copied to the temporary location. Better
solution is to leave it where it was to make access faster.

If the file is expected to be copied to temporeary location, it should
be done explicitely.