Improve the compressed file API
Distinguish between temporary storage (temp_dir) and permanent storage
(output_dir).

The temporary storage is by definition meant to be cleaned up on exit. Declaring
the output_dir will cause the uncompressed file not to be deleted after
recompression!

The permanent mode is currently used for caching of uncompressed files.